console-reader/requirements.txt
websockets
asyncio
requests
python-dotenv
